DealDamage
When the enemy reaches the end of the path, it now deals damage to the player. Removed Storpingvin (replaced by PlayerStats). EnemyModel now has variables for speed, damage and health. EnemyController has the DealDamage function in it.

+++ Penguin Tower Defense/Assets/Scrips/Lives/BaseEnemy.cs
@@ -6,5 +6,5 @@ public class BaseEnemy : MonoBehaviour
 {
     protected int damage;
     protected int health;
-    protected int speed;
+    protected float speed;
 }
